requirements.txt
aiogram
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
openai
